+++ settings/constants.go
@@ -7,7 +7,7 @@ const (
 
 	CommentWidthName        = "Comment Width"
 	CommentWidthKey         = "CLX_COMMENT_WIDTH"
-	CommentWidthDefault     = 70
+	CommentWidthDefault     = 65
 	CommentWidthDescription = "Sets the maximum number of characters on each line for comments, replies and " +
 		"descriptions in settings. Set to \u001B[1m0\u001B[0m to use the whole screen."
 	IndentSizeName        = "Indent Size"